+++ assets/js/dashboard/realtime.js
@@ -39,14 +39,14 @@ class Realtime extends React.Component {
             <Datepicker site={this.props.site} query={this.props.query} />
           </div>
         </div>
-        <VisitorGraph site={this.props.site} query={this.props.query} timer={this.props.timer} />
+        <VisitorGraph site={this.props.site} query={this.props.query} />
         <div className="items-start justify-between block w-full md:flex">
-          <Sources site={this.props.site} query={this.props.query} timer={this.props.timer} />
-          <Pages site={this.props.site} query={this.props.query} timer={this.props.timer} />
+          <Sources site={this.props.site} query={this.props.query} />
+          <Pages site={this.props.site} query={this.props.query} />
         </div>
         <div className="items-start justify-between block w-full md:flex">
-          <Locations site={this.props.site} query={this.props.query} timer={this.props.timer} />
-          <Devices site={this.props.site} query={this.props.query} timer={this.props.timer} />
+          <Locations site={this.props.site} query={this.props.query} />
+          <Devices site={this.props.site} query={this.props.query} />
         </div>
 
         { this.renderConversions() }