+++ assets/js/dashboard/stats/sources/search-terms.js
@@ -5,15 +5,21 @@ import MoreLink from '../more-link'
 import numberFormatter from '../../util/number-formatter'
 import RocketIcon from '../modals/rocket-icon'
 import * as api from '../../api'
+import LazyLoader from '../../components/lazy-loader'
 
 export default class SearchTerms extends React.Component {
   constructor(props) {
     super(props)
     this.state = {loading: true}
+    this.onVisible = this.onVisible.bind(this)
+    this.fetchSearchTerms = this.fetchSearchTerms.bind(this)
   }
 
-  componentDidMount() {
+  onVisible() {
     this.fetchSearchTerms()
+    if (this.props.query.period === 'realtime') {
+      document.addEventListener('tick', this.fetchSearchTerms)
+    }
   }
 
   componentDidUpdate(prevProps) {
@@ -23,6 +29,10 @@ export default class SearchTerms extends React.Component {
     }
   }
 
+  componentWillUnmount() {
+    document.removeEventListener('tick', this.fetchSearchTerms)
+  }
+
   fetchSearchTerms() {
     api.get(`/api/stats/${encodeURIComponent(this.props.site.domain)}/referrers/Google`, this.props.query)
       .then((res) => this.setState({
@@ -30,7 +40,7 @@ export default class SearchTerms extends React.Component {
         searchTerms: res.search_terms || [],
         notConfigured: res.not_configured,
         isAdmin: res.is_admin
-      })).catch((error) => 
+      })).catch((error) =>
         {
             this.setState({ loading: false, searchTerms: [], notConfigured: true, error: true, isAdmin: error.payload.is_admin })
         }
@@ -122,7 +132,9 @@ export default class SearchTerms extends React.Component {
       >
         { this.state.loading && <div className="loading mt-44 mx-auto"><div></div></div> }
         <FadeIn show={!this.state.loading} className="flex-grow">
-          { this.renderContent() }
+          <LazyLoader onVisible={this.onVisible}>
+            { this.renderContent() }
+          </LazyLoader>
         </FadeIn>
       </div>
     )