+++ assets/js/dashboard/util/realtime-update-timer.js
@@ -1,0 +1,8 @@
+const THIRTY_SECONDS = 30000
+const tickEvent = new Event('tick')
+
+export function start() {
+    setInterval(() => {
+        document.dispatchEvent(tickEvent)
+    }, THIRTY_SECONDS)
+}